Move `build-thread` => `util`
