+++ src/sidebar/util/thread.js
@@ -1,4 +1,4 @@
-/** @typedef {import('../build-thread').Thread} Thread */
+/** @typedef {import('../util/build-thread').Thread} Thread */
 
 /**
  * Count the number of annotations/replies in the `thread` whose `visible`
